Breaking change: drop JSON element from Event. This is a tough change to make, but there a few reasons why.
First, the Asana API returns many duplicates for any given event. Since Event is a data class, we'd like to have equivalent hashcodes for equivalent events. However, there's no simple way to accomplish that AND hold the JSON body as a property of Event objects.

Second, the JSON body is marginally useful, and only for change events.